+++ Source code/sfpca/.Rproj.user/A8A22B62/persistent-state
@@ -1,6 +1,6 @@
 build-last-errors="[]"
 build-last-errors-base-dir="~/Desktop/MoMA-prototype/Source code/sfpca/"
-build-last-outputs="[{\"output\":\"==> Rcmd.exe INSTALL --preclean --no-multiarch --with-keep.source sfpca\\n\\n\",\"type\":0},{\"output\":\"* installing to library 'C:/Users/Smart/R/win-library/3.4'\\r\\n\",\"type\":1},{\"output\":\"* installing *source* package 'sfpca' ...\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"** libs\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++  -std=gnu++11 -I\\\"C:/PROGRA~1/R/R-34~1.1/include\\\" -DNDEBUG  -I\\\"C:/Users/Smart/R/win-library/3.4/Rcpp/include\\\" -I\\\"C:/Users/Smart/R/win-library/3.4/RcppArmadillo/include\\\"   -I\\\"d:/Compiler/gcc-4.9.3/local330/include\\\"  -fopenmp    -O2 -Wall  -mtune=core2 -c RcppExports.cpp -o RcppExports.o\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++  -std=gnu++11 -I\\\"C:/PROGRA~1/R/R-34~1.1/include\\\" -DNDEBUG  -I\\\"C:/Users/Smart/R/win-library/3.4/Rcpp/include\\\" -I\\\"C:/Users/Smart/R/win-library/3.4/RcppArmadillo/include\\\"   -I\\\"d:/Compiler/gcc-4.9.3/local330/include\\\"  -fopenmp    -O2 -Wall  -mtune=core2 -c model.cpp -o model.o\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++  -std=gnu++11 -I\\\"C:/PROGRA~1/R/R-34~1.1/include\\\" -DNDEBUG  -I\\\"C:/Users/Smart/R/win-library/3.4/Rcpp/include\\\" -I\\\"C:/Users/Smart/R/win-library/3.4/RcppArmadillo/include\\\"   -I\\\"d:/Compiler/gcc-4.9.3/local330/include\\\"  -fopenmp    -O2 -Wall  -mtune=core2 -c sfpca_extended.cpp -o sfpca_extended.o\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++ -shared -s -static-libgcc -o sfpca.dll tmp.def RcppExports.o model.o sfpca_extended.o -fopenmp -LC:/PROGRA~1/R/R-34~1.1/bin/x64 -lRlapack -LC:/PROGRA~1/R/R-34~1.1/bin/x64 -lRblas -lgfortran -lm -lquadmath -Ld:/Compiler/gcc-4.9.3/local330/lib/x64 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/PROGRA~1/R/R-34~1.1/bin/x64 -lR\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"installing to C:/Users/Smart/R/win-library/3.4/sfpca/libs/x64\\r\\n\",\"type\":1},{\"output\":\"** R\\r\\n\",\"type\":1},{\"output\":\"** preparing package for lazy loading\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"** help\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"Warning: C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/man/sfpca-package.Rd:27: All text must be in a section\\r\\n\",\"type\":2},{\"output\":\"Warning: C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/man/sfpca-package.Rd:28: All text must be in a section\\r\\n\",\"type\":2},{\"output\":\"Warning: C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/man/sfpca-package.Rd:29: All text must be in a section\\r\\n\",\"type\":2},{\"output\":\"*** installing help indices\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"** building package indices\\r\\n\",\"type\":1},{\"output\":\"** testing if installed package can be loaded\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"* DONE (sfpca)\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1}]"
+build-last-outputs="[{\"output\":\"==> Rcmd.exe INSTALL --preclean --no-multiarch --with-keep.source sfpca\\n\\n\",\"type\":0},{\"output\":\"* installing to library 'C:/Users/Smart/R/win-library/3.4'\\r\\n\",\"type\":1},{\"output\":\"* installing *source* package 'sfpca' ...\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"** libs\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++  -std=gnu++11 -I\\\"C:/PROGRA~1/R/R-34~1.1/include\\\" -DNDEBUG  -I\\\"C:/Users/Smart/R/win-library/3.4/Rcpp/include\\\" -I\\\"C:/Users/Smart/R/win-library/3.4/RcppArmadillo/include\\\"   -I\\\"d:/Compiler/gcc-4.9.3/local330/include\\\"  -fopenmp    -O2 -Wall  -mtune=core2 -c RcppExports.cpp -o RcppExports.o\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++  -std=gnu++11 -I\\\"C:/PROGRA~1/R/R-34~1.1/include\\\" -DNDEBUG  -I\\\"C:/Users/Smart/R/win-library/3.4/Rcpp/include\\\" -I\\\"C:/Users/Smart/R/win-library/3.4/RcppArmadillo/include\\\"   -I\\\"d:/Compiler/gcc-4.9.3/local330/include\\\"  -fopenmp    -O2 -Wall  -mtune=core2 -c model.cpp -o model.o\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++  -std=gnu++11 -I\\\"C:/PROGRA~1/R/R-34~1.1/include\\\" -DNDEBUG  -I\\\"C:/Users/Smart/R/win-library/3.4/Rcpp/include\\\" -I\\\"C:/Users/Smart/R/win-library/3.4/RcppArmadillo/include\\\"   -I\\\"d:/Compiler/gcc-4.9.3/local330/include\\\"  -fopenmp    -O2 -Wall  -mtune=core2 -c sfpca_extended.cpp -o sfpca_extended.o\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++  -std=gnu++11 -I\\\"C:/PROGRA~1/R/R-34~1.1/include\\\" -DNDEBUG  -I\\\"C:/Users/Smart/R/win-library/3.4/Rcpp/include\\\" -I\\\"C:/Users/Smart/R/win-library/3.4/RcppArmadillo/include\\\"   -I\\\"d:/Compiler/gcc-4.9.3/local330/include\\\"  -fopenmp    -O2 -Wall  -mtune=core2 -c solver.cpp -o solver.o\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"c:/Rtools/mingw_64/bin/g++ -shared -s -static-libgcc -o sfpca.dll tmp.def RcppExports.o model.o sfpca_extended.o solver.o -fopenmp -LC:/PROGRA~1/R/R-34~1.1/bin/x64 -lRlapack -LC:/PROGRA~1/R/R-34~1.1/bin/x64 -lRblas -lgfortran -lm -lquadmath -Ld:/Compiler/gcc-4.9.3/local330/lib/x64 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/PROGRA~1/R/R-34~1.1/bin/x64 -lR\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"installing to C:/Users/Smart/R/win-library/3.4/sfpca/libs/x64\\r\\n\",\"type\":1},{\"output\":\"** R\\r\\n\",\"type\":1},{\"output\":\"** preparing package for lazy loading\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"** help\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"Warning: C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/man/sfpca-package.Rd:27: All text must be in a section\\r\\n\",\"type\":2},{\"output\":\"Warning: C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/man/sfpca-package.Rd:28: All text must be in a section\\r\\n\",\"type\":2},{\"output\":\"Warning: C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/man/sfpca-package.Rd:29: All text must be in a section\\r\\n\",\"type\":2},{\"output\":\"\",\"type\":1},{\"output\":\"*** installing help indices\\r\\n\",\"type\":1},{\"output\":\"** building package indices\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"** testing if installed package can be loaded\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1},{\"output\":\"* DONE (sfpca)\\r\\n\",\"type\":1},{\"output\":\"\",\"type\":1}]"
 compile_pdf_state="{\"errors\":[],\"output\":\"\",\"running\":false,\"tab_visible\":false,\"target_file\":\"\"}"
 console_procs="[]"
 files.monitored-path=""
