+++ Source code/sfpca/src/solver.cpp
@@ -1,0 +1,6 @@
+// #include "solver.h"
+
+// arma::vec Grad_nosmoothv(arma::vec x)
+// {
+//     return -VCONST_VEC + x;
+// }
